Test to see if directory is deleable. This isn't perfect (it could be but stuff in it not), but should give a rough idea if the delete will work. If not deletable, raise an exception (which should kill the delayed job and let us recover once we see those going bad).

+++ app/models/job/ingest_staging_delete.rb
@@ -19,7 +19,11 @@ class Job::IngestStagingDelete < Job::Base
 
   def perform
     if File.directory?(self.path)
-      FileUtils.rm_rf(self.path)
+      if File.writable?(self.path)
+        FileUtils.rm_rf(self.path)
+      else
+        raise RuntimeError, "IngestStagingDelete: Medusa does not have permission to delete #{self.path}."
+      end
     end
     if self.external_file_group.present?
       Workflow::IngestMailer.staging_delete_done(self.user, self.external_file_group).deliver_now
